requirements.txt
Flask==2.3.2
google-cloud-storage==2.9.0
google-cloud-firestore==2.11.0
tensorflow==2.13.0
Pillow==10.0.0
numpy==1.24.2
pytz==2023.3